requirements.txt
streamlit
sentence-transformers
google-generativeai
sentence_transformers 
google.colab 
pandas
openpyxl
torch
numpy